pyproject.toml
[project]
name = "cxi-pipeline-ray"
version = "0.1.0"
description = "Ray-based CPU post-processing pipeline for PeakNet inference results"
authors = [
    {name = "Chun Hong Yoon", email = "cxyoon@slac.stanford.edu"},
]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    "ray>=2.0.0",
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "h5py>=3.0.0",
    "torch>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
cxi-writer = "cxi_pipeline_ray.cli:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["cxi_pipeline_ray*"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"
